+++ doc/source/doxygen-docs/changelog.md
@@ -10,6 +10,7 @@
     - Address the new field `variance` in `sensor_msgs/Range` (Closes [issue #1270](https://github.com/MRPT/mrpt/issues/1270)).
 - BUG FIXES:
   - Fix wrong rendering of all wxWidgets-based OpenGL windows when using Ubuntu's display settings to change UI to a size different than 100% (Fixes [issue #1114](https://github.com/MRPT/mrpt/issues/1114)).
+  - Fix ignored sensorPose of mrpt::obs::CObservationPointCloud while inserting them into voxel maps.
 
 # Version 2.11.1: Released Oct 23rd, 2023
 - Changes in libraries: